feat(task-service): rename sub task to user task interface
BB-1785
