+++ services/task-service/src/interfaces/index.ts
@@ -4,4 +4,4 @@ export * from './i-event-adapter';
 export * from './i-event-processor';
 export * from './i-incoming-connector';
 export * from './i-outgoing-connector';
-export * from './i-sub-task-service';
+export * from './i-user-task-service';
